refactor(type_checker): extract methods/struct_impl.rs
