+++ src/type_checker/methods/mod.rs
@@ -1,5 +1,6 @@
 mod array;
 mod string;
+mod struct_impl;
 mod tcp;
 
 use super::{Type, TypeChecker};
@@ -82,57 +83,8 @@ impl TypeChecker {
             string::check(self, Type::String, method_name, args)
         } else if matches!(object_type, Type::TcpListener | Type::TcpStream) {
             tcp::check(self, object_type, method_name, args)
-        } else if let Type::Struct(struct_name) = object_type {
-            let (return_type, param_types) = match self.struct_methods.get(&struct_name) {
-                Some(methods) => match methods.get(method_name) {
-                    Some((ret, params)) => (ret.clone(), params.clone()),
-                    None => {
-                        let available: Vec<String> = methods.keys().cloned().collect();
-                        return Err(format!(
-                            "Method '{}' not found on struct '{}'\n   = help: Available methods: {}",
-                            method_name,
-                            struct_name,
-                            available.join(", ")
-                        ));
-                    }
-                },
-                None => {
-                    return Err(format!(
-                        "Struct '{}' has no methods defined\n   = help: Add methods with 'impl {} {{ fun method(self, ...) {{ ... }} }}'",
-                        struct_name, struct_name
-                    ));
-                }
-            };
-            if args.len() + 1 != param_types.len() {
-                return Err(format!(
-                    "Method '{}' on '{}' expects {} arguments (including self), got {}\n   = help: Expected signature: {}({})",
-                    method_name,
-                    struct_name,
-                    param_types.len() - 1,
-                    args.len(),
-                    method_name,
-                    param_types
-                        .iter()
-                        .skip(1)
-                        .map(|t| t.fmt_for_user())
-                        .collect::<Vec<_>>()
-                        .join(", ")
-                ));
-            }
-            for (i, arg) in args.iter().enumerate() {
-                let arg_type = self.check_expression(arg)?;
-                let expected = &param_types[i + 1];
-                if arg_type != *expected {
-                    return Err(format!(
-                        "Method '{}' argument {} expects {}, got {}",
-                        method_name,
-                        i + 1,
-                        expected.fmt_for_user(),
-                        arg_type.fmt_for_user()
-                    ));
-                }
-            }
-            Ok(return_type)
+        } else if let Type::Struct(_) = object_type {
+            struct_impl::check(self, object_type, method_name, args)
         } else {
             Err(format!(
                 "Cannot call method on value of type '{}'\n   = help: Methods work on arrays, strings, modules, and structs with impl blocks.",
